requirements.txt
asgiref==3.8.1
Django==4.2.21
django-debug-toolbar==5.2.0
django-postgres-extra==2.0.8
django-rest-framework==0.1.0
djangorestframework==3.16.0
Faker==37.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
 